Projetos-Flask/ChatEat/requirements.txt
Flask
python-dotenv
PyMySQL
argon2-cffi
abacatepay
qrcode[pil]
